+++ BTrees/intvaluemacros.h
@@ -31,7 +31,7 @@
       long vcopy = INT_AS_LONG(ARG);                            \
       if (PyErr_Occurred()) { (STATUS)=0; (TARGET)=0; }           \
       else if ((int)vcopy != vcopy) {                                  \
-        PyErr_SetString(PyExc_OverflowError, "integer out of range"); \
+        PyErr_SetString(PyExc_TypeError, "integer out of range"); \
         (STATUS)=0; (TARGET)=0;                                   \
       }                                                           \
       else TARGET = vcopy;                                        \
@@ -64,11 +64,11 @@
       long vcopy = INT_AS_LONG(ARG);                              \
       if (PyErr_Occurred()) { (STATUS)=0; (TARGET)=0; }           \
       else if (vcopy < 0) {                                       \
-        PyErr_SetString(PyExc_OverflowError, "can't convert negative value to unsigned int"); \
+        PyErr_SetString(PyExc_TypeError, "can't convert negative value to unsigned int"); \
         (STATUS)=0; (TARGET)=0;                                   \
-      }								  \
+      }                                                           \
       else if ((unsigned int)vcopy != vcopy) {                    \
-        PyErr_SetString(PyExc_OverflowError, "integer out of range"); \
+        PyErr_SetString(PyExc_TypeError, "integer out of range"); \
         (STATUS)=0; (TARGET)=0;                                   \
       }                                                           \
       else TARGET = vcopy;                                        \